// src/types/types.ts
export type IdParam = string | undefined | number;
// ...
export enum TicketStatus {
  Active = 'ACTIVE',
  Inactive = 'INACTIVE',
}

export interface TicketPrice {
  id?: number;
  label?: string;
  price: number;
  sale_start_date?: string | Date;
  sale_end_date?: string | Date;
  price_including_taxes_and_fees?: number;
  price_before_discount?: number;
  is_discounted?: boolean;
  tax_total?: number;
  fee_total?: number;
  is_available?: boolean;
  is_before_sale_start_date?: boolean;
  is_after_sale_end_date?: boolean;
  is_sold_out?: boolean;
  initial_quantity_available?: number;
  quantity_sold?: number;
  isHidden?: boolean;
  quantity_remaining?: number;
}

export interface Ticket {
  id?: number;
  order?: number;
  title: string;
  event_id?: IdParam;
  type: Ticket;
  description?: string;
  price?: number;
  prices?: TicketPrice[];
  price_before_discount?: number;
  is_discounted?: boolean;
  initial_quantity_available?: number | undefined;
  quantity_sold?: number;
  sale_start_date?: string | Date;
  sale_end_date?: string | Date;
  max_per_order?: number;
  min_per_order?: number;
  hide_before_sale_start_date?: boolean;
  hide_after_sale_end_date?: boolean;
  hide_when_sold_out?: boolean;
  start_collapsed?: boolean;
  show_quantity_remaining?: boolean;
  quantity_available?: number;
  status?: TicketStatus;
  is_sold_out?: boolean;
  is_available?: boolean;
  isHidden_without_promo_code?: boolean;
  is_before_sale_start_date?: boolean;
  is_after_sale_end_date?: boolean;
  taxes?: TaxAndFee[];
  tax_total?: number;
  service_fee_total?: number;
  price_including_taxes_and_fees?: number;
  tax_and_fee_ids?: IdParam[];
  taxes_and_fees?: TaxAndFee[];
  isHidden?: boolean;
}
// ...
export enum QuestionType {
  ADDRESS = 'ADDRESS',
  SINGLE_LINE_TEXT = 'SINGLE_LINE_TEXT',
  MULTI_LINE_TEXT = 'MULTI_LINE_TEXT',
  CHECKBOX = 'CHECKBOX',
  RADIO = 'RADIO',
  DROPDOWN = 'DROPDOWN',
  DATE = 'DATE',
}
// ...
export enum TaxAndFeeType {
  Tax = 'TAX',
  Fee = 'FEE',
}

export enum TaxAndFeeCalculationType {
  Percentage = 'PERCENTAGE',
  Fixed = 'FIXED',
}

export interface TaxAndFee {
  id?: number;
  name: string;
  rate: number | undefined;
  type: TaxAndFeeType;
  calculation_type: TaxAndFeeCalculationType;
  is_default: boolean;
  is_active: boolean;
  description?: string;
  account_id?: IdParam;
}